requirements.txt
streamlit
Pillow==8.3.1
torch==1.7.1
torchvision==0.8.2
